--- writings/its-the-right-way.md
@@ -1,62 +1,0 @@
----
-title: "It's the right way."
-date: "2020-08-17"
-og:
-  description: "The right way of how to get into coding."
-  image: "https://telmo.im/og/right-way.png"
-author:
-  twitter: "telmo"
-  name: "Telmo Goncalves"
----
-
-I don't want to waste your time with this article so I'll say it upfront,
-**the right way of getting into coding doesn't exist.**
-
-Now that we get that out of the way lets get started. This is highly focused on
-my own experience, I started coding in 2007.
-
----
-
-# How to get into coding
-
-If you've been on Twitter I bet you've seen accounts bullshitting their way into
-_teaching you the right way to code_, BS. They _say_ after HTML + CSS
-you **should** learn JavaScript basics before jumping into libraries.
-
-When I started learning how to code I had no idea what the hell I was doing,
-as most of us. I started learning HTML + CSS, the concept of a `div` being a box
-was too much for me, it's perfectly normal thinking and questioning every single thing
-about coding.
-
-Now per almost everyone I **had** to jump into JavaScript basics, well, I didn't. I went
-straight from HTML + CSS into PHP, and you know what? It's fine. Back in the day, I was
-using PHP for frontend development. I've been called a noob and that _"I don't know how to code at all"_
-because I used PHP for frontend, if you're one of those folks I can honestly say — screw you.
-
-Back to JavaScript, again, before learning the basics I started learning jQuery straight away,
-the basics came later, and once again it was fine.
-
----
-
-# Your way is the right way
-
-I agree if you're working with a team you **should** follow best practices, but again, each
-team has its own way of working, you'd need to adapt to their way from project to project.
-
-What happened to me by not _learning things the **right** way_ was that I'm curious,
-eager to learn and most importantly I keep on building stuff, even after 13+ years.
-
-If you want to learn PHP or Ruby on Rails just do it. It's OK to follow some advice here and there,
-but sometimes it's just too much, it feels like _experienced_ folks are putting too much
-pressure on you to follow the right way.
-
-Sometime we're taking advice from folks that never built a single application in their
-life, now tell me, if they've never built an application how the hell are they in a position
-of telling others how they should learn how to code? Spoiler alert, **they're not**.
-
----
-
-If you want to know what I've learned over the years check out my [about page](/about).
-
-<br />
-<br />
