--- writings/my-coding-journey.md
@@ -1,100 +1,0 @@
----
-title: "How I ended up coding"
-date: "2020-07-02"
-og:
-  description: "How I ended up coding for a living. Ups and Downs!"
-  image: "https://telmo.im/og/howendedupcoding.png"
-author:
-  twitter: "telmo"
-  name: "Telmo Goncalves"
----
-
-_This is my journey on how I got into code, everyone has a different one, this one is mine._
-
-<br>
-
-# Early years
-
-Before I dive into telling you how I ended up coding I would like to tell you what
-I was up to before that.
-
-When I was 15 years old my parents got me my first computer, I could've used it for
-coding, but no. That computer purpose was for mIRC, play video games and watch movies.
-
-My first interaction with code was due to a group I had with my friends, one summer day we
-were talking that it would be cool to have a website of the group, I immediately dived into
-figure out how to build a website.
-
-I got as far as using a `<h1>` for a title and a `background-image` that for the life of me
-I couldn't figure out how to prevent it to repeat itself.
-
-You guessed it, `background-repeat: no-repeat`. I quickly dropped that idea and just
-thought that this was definitely not for me.
-
----
-
-# Before coding
-
-I dropped out of school in the 10th grade. I did half a year of arts, I developed a love for
-design since I was young. After that I did half a year in economics, attending school at night
-and living my teenage years during the day. What a waste of a year, geez.
-
-When I was 17 years old I started working in a bar, pouring drinks and making sandwiches, it was not bad, I was living with my parents, worry-free. I spent around 1 year working there.
-
-At 18 I decided to get a more _serious_ job so I started working in a warehouse, picking products
-into palettes for the major supermarkets in Portugal, I spent 2 years working there, and to be honest some
-days I do miss it.
-
-While in the warehouse I started learning how to build layouts with Photoshop and slowly converting those
-into HTML & CSS. We _(a friend and I)_ got a job for a client, building a website for €4000, that's when it clicked, it was a lot of money, we could do this full-time.
-
----
-
-# One startup
-
-At 20 years old when I quit my warehouse job I decided to try my luck by creating a startup with my
-friend. He quit his job as well, and we went for it.
-
-We'd created our company in May 2008 and couldn't get a single client until September that year,
-we kept scratching our heads trying to figure out what the hell we were doing wrong but once we got that
-first client, things started to flow naturally.
-
-By the end of 2009, we had settled a big contract with a company in London, we were travelling back
-and forth, having the time of our lives.
-
-Things in Portugal were not going as we expected, the economy was crashing, taxes were going up and by the end of 2012 (when the London contract ended), we decided to close the startup and be done with it.
-
-I mean, it was good while it lasted, we learned a LOT from it, but it was time.
-
----
-
-# First tech job
-
-In September 2012 I moved to London looking for a better job, landed my first job after a month. I had
-plans to stay in London for at least 2 years and no longer than 5 years, I ended up staying there only
-for 5 months.
-
-My girlfriend was back in Portugal, I started missing her, my friends, family, everything to be
-honest. I ended up making a deal with my employee, I could do a 3 month trial on remote work.
-
-I ended up working for them for 7 years while remote, with the occasional trip to London. In there
-I had the chance to learn quite a lot, that's where I learned Angular, TypeScript, ReactJS, RoR, etc;
-
----
-
-# So far
-
-I'm now working for a great company full of talented people and an amazing culture. I live happily
-in Portugal with my girlfriend, our black labrador and expecting our first newborn.
-
-Cheers to that! 🥂
-
----
-
-# What else?
-
-What else can I say? Well, if you're looking for a job don't be afraid to apply, some people
-don't apply just because they don't match 1 item on the requirements list, instead of not applying, see it this way: _"Alright, I match 99% of the requirements this company is asking."_ — go for it.
-
-With my startup, I learned a few things as well, and the most important one was how
-to say **"No!"**, sometimes you just need to draw the line.
