--- writings/no-you-dont-suck.md
@@ -1,62 +1,0 @@
----
-title: "No, you don't suck!"
-date: "2020-04-24"
-og:
-  description: "Developers don't know everything, and that's OK."
-  image: "https://telmo.im/og/youdontsuck.png"
-author:
-  twitter: "telmo"
-  name: "Telmo Goncalves"
----
-
-I've been noticing that one of the biggest issues with new developers is that they think they need to know **EVERYTHING**.
-
-Don't fool yourself, there's not one single developer in the world that knows everything, all those high tech people you look up to, they **don't** know everything, that's a fact. You can always argue with someone about their opinion, you can't argue against facts.
-
-I've been coding for 13+ years, you might think "Wow, that's a lot, he must know so much." — well, yeah, it's a lot but I guarantee you I don't know everything, I'm far from it. But, the good news it that it's OK, we don't need to know everything.
-
----
-
-Let me give you an example. You might ask me _how to find an odd number in JavaScript_, do I know how to do it? No, I don't, but here's some code:
-
-```js
-function isOdd(num) { return num % 2 }
-```
-
-Now, you might think, he **did** knew how to do it. Again, no, I didn't, I just googled it. Am I cheating at programming? **No!**
-
-_Well, you googled it, sure, but you know what `%` it's doing_. Flash news, no, I don't. I have absolutely no idea what it's doing.
-
-And again, that's OK.
-
----
-
-Want another example? Sure.
-
-_"Hey, you know how to use `useState` in React?"_
-
-```js
-const [published, setPublished] = useState(false);
-
-// [...]
-
-setPublished(true)
-
-// [...]
-
-published ? "Well, yes!" : "Not yet, sorry!"
-```
-
-## How useState works?
-
-No fucking idea, I know it updates a component's state, I know how to set it, to change it, when to use it, etc; What's happening behind the scenes? 🤷‍♂️
-
----
-
-Most developers don't know a LOT of things, they just know how to un-shit themselves.
-
-I know how to build things, I learn fast, but if you ask me what's a `function` I have no idea, I can explain what it does, I can explain how and when to use it. But the technical description of _"A JavaScript function is a block of code designed to perform a particular task."_, fuck that.
-
----
-
-I had to write this, I get a little frustrated when a lot of people think they must know everything. **You don't.**
