--- writings/open-sourcing-blog.md
@@ -1,39 +1,0 @@
----
-title: "I'm open sourcing my website"
-date: "2020-04-30"
-og:
-  description: "I decided to open source my website as it is, no cleaning up."
-  image: "https://telmo.im/og/ooswebsite.png"
-author:
-  twitter: "telmo"
-  name: "Telmo Goncalves"
----
-
-I've decided to open source this website. When I first launched it I've posted on Twitter about it asking for feedback. One of the comments was from [Catalin](https://twitter.com/catalinmpit):
-
-<blockquote class="twitter-tweet"><p lang="en" dir="ltr">Looks awesome. Teach me.</p>&mdash; Catalin Pit 🤷‍♂️ (@catalinmpit) <a href="https://twitter.com/catalinmpit/status/1252912096528007170?ref_src=twsrc%5Etfw">April 22, 2020</a></blockquote> <script async src="https://platform.twitter.com/widgets.js" charset="utf-8"></script>
-
-This gave me the idea of open sourcing this code, thanks Catalin.
-
----
-
-This is a really short article, I just want to explain why I'm open sourcing this code. First of all, I saw a lot of interest from comments on Twitter, feel free to fork, clone, whatever this code and use it for your own blog.
-
-Second, I thought about tidy up the code, putting it all pretty and awesome, so when people would check it they would be all like _"Holy crap! This code is sooo organized, this is so awesome!"_, screw that, seriously!
-
-By not doing that you can see that I also write messy code, I'm mostly talking about `base.css`, but feel free to judge all the other files as well.
-
-This can show that it's perfectly normal to write messy code, you might look at my website and think that it looks great, so clean, minimalist design, etc; But you don't know how the code looks like, well, now you can check it out.
-
----
-
-## Repo
-
-I'll post the repo here, if you're one of those persons that just like to contribute to whatever project you get your hand on, feel free to open pull requests.
-
-If you want to use this code to build your own website, blog, etc; Go for it.
-
-[Open Github](https://github.com/telmogoncalves/telmo)
-
-
-
